Implementing route protection with JWT Auth

+++ utils/jwt.go
@@ -45,6 +45,6 @@ func VerifyToken(token string) (int64, error) {
 	}
 
 	//email := claims["email"].(string)
-	userId := claims["userId"].(int64)
+	userId := int64(claims["userId"].(float64))
 	return userId, nil
 }
